added multiboot2 and uefi (1/2) functionality
Multiboot2 information is now properly parsed and put into place.
Inside the boot module, there are now new files, `_multiboot2.rs`
and `_uefi.rs` which handle multiboot2 and UEFI respectively. The
UEFI functionality (i.e. exiting the boot services) is not yet
complete.

We are lacking a `Handle` which ought to be provided to
`kernel_main`? This needs some investigation.