adjusted log level for messages, format and build script
Certain messages got their log level adjusted. The logging format
for the QEMU debug logger was altered (more width for the file that
caused the log message).

The GitHub workflow was adjusted to use the correct toolchain now
(i.e. the one that is also found in `kernel/rust-toolchain.toml`).

The build script was adjusted for better readability. Kernel variables
(used during build time) are now using `export` because with `export`,
the need not be written again in front of the build command.
    If the kernel binary is not multiboot2 compatible, an error message
is now shown that indicates that something is off. The build will not
continue anymore.

The `_uefi` module now has a statically allocated array that _should_
be able to hold the boot services memory map information. This is
checked with an `assert!`.