relocated assembly boot code
With upcoming commits, hardware specific initialization takes place
and it is beneficial to have all architecture dependent code in one
place. The assembly boot code (architecture dependent code) was
moved to `kernel/library/hardware/{architecture}/boot`. The `boot`
module in the `library` was kept as its code is architecture inde-
pendent. The paths to the assembly code and to the linker (for the
target definition) were adjusted.