Proper `mutiboot2` Implementation (#15)
* added kernel stack overflow integration test

* started working on (virtual) memory

* got page table working with bootloader

* major linting enhancements

* updated tests and resolved clippy errors

* added generic boot information, .gitattributes and better test documentation

* reformatted .gitattributes

* renamed and refactored the boot workspace member

The `boot` crate is now called `init`. Additionally, conditional
compilation was introduced to be able to compile and run the kernel
for different architectures.

* reworked boot information and architecture handling procedure

Each architecture now has its own entrypoint to initialize architecture
specific systems (again). This is not the most DRY approach, but the
most flexible and the easiest to implement, as we can rely on
conditional compilation.

* added `prelude` module to be imported everywhere

The `prelude` module contains often used functions and modules and can
be imported with `use crate::prelude::*`

* resolved linting errors

* completely change to multiboot2

* add license template and added license note to source code

* cleanup

* more (dependency) cleanup, refactoring and code sanitizing

the `hardware` module is now located under `build/old/` as it has become outdated

* updated `x86_64` dependency and Rust toolchain

With the new version (0.14.7) of the `x86_64` crate, we get the latest
nightly toolchain and therefore Rust 1.59. With Rust 1.59, inline
assembly was stabilized and we can omit the feature gates.

* introduce refactored logging

The new logger uses the logging facade provided by the `log` crate
and makes uses of the `ansi_rgb` crate to color the output.

* interim commit

* fixed assembly boot code

The GAS (GNU AT&T assembly dialect) boot code had minor errors:

1. The `multiboot2_header.S` file disabled the UEFI Boot Services
   temporarily, which were re-enabled. Otherwise, KVM will throw
   a nasty error.
2. The `start.S` file had multiple issues. Some register names were
   wrong, 32bit registers were mixed up with 64bit ones. Therefore,
   the correct suffixes for each instruction were added. Moreover,
   the stack size was increased (x10).

* improved build and run scripts

Added a new command to the `clean` recipe in the Justfile and re-
factored the scripts.

* formatting and cleanup

* added compilation build and time

The kernel can now log the build time and date during startup.

* refactored log

The structure of the `log` module now makes a lot more sense. The
serial port got its own structure and the `KernelLog` now "only"
forwards messages to the loggers if they are enabled.

* added multiboot2 and uefi (1/2) functionality

Multiboot2 information is now properly parsed and put into place.
Inside the boot module, there are now new files, `_multiboot2.rs`
and `_uefi.rs` which handle multiboot2 and UEFI respectively. The
UEFI functionality (i.e. exiting the boot services) is not yet
complete.

We are lacking a `Handle` which ought to be provided to
`kernel_main`? This needs some investigation.

* adjusted log level for messages, format and build script

Certain messages got their log level adjusted. The logging format
for the QEMU debug logger was altered (more width for the file that
caused the log message).

The GitHub workflow was adjusted to use the correct toolchain now
(i.e. the one that is also found in `kernel/rust-toolchain.toml`).

The build script was adjusted for better readability. Kernel variables
(used during build time) are now using `export` because with `export`,
the need not be written again in front of the build command.
    If the kernel binary is not multiboot2 compatible, an error message
is now shown that indicates that something is off. The build will not
continue anymore.

The `_uefi` module now has a statically allocated array that _should_
be able to hold the boot services memory map information. This is
checked with an `assert!`.

* removed `kernel/old/` directory

* relocated assembly boot code

With upcoming commits, hardware specific initialization takes place
and it is beneficial to have all architecture dependent code in one
place. The assembly boot code (architecture dependent code) was
moved to `kernel/library/hardware/{architecture}/boot`. The `boot`
module in the `library` was kept as its code is architecture inde-
pendent. The paths to the assembly code and to the linker (for the
target definition) were adjusted.

* miscellaneous small adjustments

* relocated old x86_64 hardware initialization code

The source files previously found under `kernel/old/` were now
put into their correct location. These files ARE NOT ready to be
used - they were placed there to make future development easier
and current design decisions (about the directory structure) more
comprehensible.

The `hardware` module now contains more code, but not much of this
code is used as of now.

* really put all arch-dependent code into `hardware`

Not all architecture dependent code was located under
`library/hardware`. This changed now.

* updated documentation and `tools` script

The documentation was updated to be up-to-date with the source
code. There were some additions as well, mainly when it comes
to CI/CD, modules and development.

The `tools.sh` script will now check for the `just` command when
trying to install `just`. `cargo-xbuild` was removed from build
dependencies.

* introduces new init.sh shell scripts

The shell scripts under `scripts/` can now use uniform shell script
located under `scripts/lib/` named `init.sh` that performas proper
initialization (setting variables and chaning directories).
    This script, if supplied with the 'kernel' argument in `$1`,
will also change into the kernel directory and perform more
initialiuation for the kernel.

* further refined scripts

A new test-wrapper script was created and functionality of the
`Justfile` was outsourced into the `init.sh` script to provide
a uniform setup experience for all scripts.
    This commit massively simplifies the `Justfile` and further
refactors the scripts to use consistent variabled.

* small documentation enhancements

* renamed the `hardware` module -> `architectures`

* exiting boot services (1/2)

The UEFI boot services can be exited as of now and we get the
memory map - everything works as expected. We are currently
waiting for a PR to get merged in the UEFI crate to make
obtaining the memory map idiomatic.

* revised scripts again

The new revised scripts improve mainly when it comes to ease of use
(especially when it comes to manual invocation). The GitHub work-
flows now use the scripts when running checks and tests for unCORE.
    The `Justfile` and the documentaton were adjusted accordingly.
Minor improvements were incorporated in addition to some minor re-
factoring.

* make scripts invocation-uniform and add usage information

* adjust `Justfile` to work with new script invocation

* last script adjustments

The kernel testing script was revised and now takes arguments
properly. The script log level 'ABORT' was removed - old abort
messages are now error messages.

* use proper `from_ptr` in `uefi` module

This commit wraps up PR #15 and therefore the multiboot2
integration. The `uefi` dependency is currently set to a custom
repository that has the proper `from_ptr` methods enabled. This
will be changed back to the official `uefi` crate later when it
has been merge upstream.