switched to new build method
Instead of using multiboot2, we no compile for the `x86_64-unknown-uefi`
target and have a new, well-defined entry-point: `efi_main`.
    Compilation and script complexity has been dramatically reduced
which in turn yields shorter scripts and more well-defined behavior. We
got rid of linking the kernel on our own, we got rid of the multiboot2
header files and the assembly code.