Rework Targets to use UEFI (#20)
* switched to new build method

Instead of using multiboot2, we no compile for the `x86_64-unknown-uefi`
target and have a new, well-defined entry-point: `efi_main`.
    Compilation and script complexity has been dramatically reduced
which in turn yields shorter scripts and more well-defined behavior. We
got rid of linking the kernel on our own, we got rid of the multiboot2
header files and the assembly code.

* adjusted the test runner

* adjusted integration tests

* cleaned up the `kernel/build/` directory

* removed old (assembly and multiboot2) boot code

* adjusted whole kernel entry

Now, `main.rs` as well as `lib.rs` have proper entrypoints and they
work (unit- and integration tests for basic booting pass).

* updated documentation for new UEFI target specification

* rounding things up

This commit fixes smaller issues and SOLVES the issue with the
`stack_overflow` integration test. I belive there was a linker issue
that prevented this test from passing.

* adjusted GitHub workflows to run on all branches

The '*' syntax excludes '/' in branch names, so we now use '**'.

* documentaion corrections & license template relocation

* proper submodules for architectures module

With the new UEFI targets, we have three architectures to choose
from currently:

1. aarch64 (ARM 64 Bit)
2. i686    (x86 32 Bit)
3. x86_64  (x86 64 Bit)

These modules are now present under `library/architectures/` to
be complete in this regard.

* compresseed the boot module into one file

In the future (i.e. TODO), when an allocator is present, we can
use `Arc<Mutex<TYPE>>` for proper multi-threading safe operation.

* final adjustments before merge