Re-Introduce x86_64 CPU Setup (#19)
* adjusted log to be (color) coherent

The Bash, kernel and test logger are now color coherent too.

* introduced CPU exception code

This commit introduces the proper CPU setup code, that is:

1. GDT - Global Descriptor Table
2. TSS - Task State Segment
3. IDT - Interrupt Descriptor Table

These tables are now properly loaded. Currently, two CPU exceptions
are properly implemented (breakpoint and double fault). The double
fault handler uses a dedicated kernel fallback stack, so we are able
to handle kernel stack overflows as well. The boot code was adjusted
accordingly.

* miscellaneous small additions and cleanup

* introduces interrupts module and made clippy happy

This commit introduces the `interrupts.rs` file, but it does not
yet implement interrupts. In the future, an APIC should handle
these interrupts.
    Clippy was also satified.

* adjusted exit behavior of runs and tests

The tests and runs now display a coherent behavior when it comes
to exiting properly. The exit code is first processed in
`run_in_qemu.sh` and then individually processes in the test
runner or the `Justfile`.

* adjusted `stack_overflow` test

The `stack_overflow` test does not work at the moment - who know
why? This commit add's the CPU initialization routine to `lib.rs`
as well.

* Rework Targets to use UEFI (#20)

* switched to new build method

Instead of using multiboot2, we no compile for the `x86_64-unknown-uefi`
target and have a new, well-defined entry-point: `efi_main`.
    Compilation and script complexity has been dramatically reduced
which in turn yields shorter scripts and more well-defined behavior. We
got rid of linking the kernel on our own, we got rid of the multiboot2
header files and the assembly code.

* adjusted the test runner

* adjusted integration tests

* cleaned up the `kernel/build/` directory

* removed old (assembly and multiboot2) boot code

* adjusted whole kernel entry

Now, `main.rs` as well as `lib.rs` have proper entrypoints and they
work (unit- and integration tests for basic booting pass).

* updated documentation for new UEFI target specification

* rounding things up

This commit fixes smaller issues and SOLVES the issue with the
`stack_overflow` integration test. I belive there was a linker issue
that prevented this test from passing.

* adjusted GitHub workflows to run on all branches

The '*' syntax excludes '/' in branch names, so we now use '**'.

* documentaion corrections & license template relocation

* proper submodules for architectures module

With the new UEFI targets, we have three architectures to choose
from currently:

1. aarch64 (ARM 64 Bit)
2. i686    (x86 32 Bit)
3. x86_64  (x86 64 Bit)

These modules are now present under `library/architectures/` to
be complete in this regard.

* compresseed the boot module into one file

In the future (i.e. TODO), when an allocator is present, we can
use `Arc<Mutex<TYPE>>` for proper multi-threading safe operation.

* final adjustments before merge

* renamed `main` to `kernel_main`

* added all x86_64 CPU exceptions

Now, all CPU exceptions are caught by a properly hooked up
handler function that prints information about the exception. Some
of these handler may not be complete, or lack functionality. This
functionality is to be implemented "on-the-fly" while these
handlers are needed.