feature: simple kernel heap allocator (#24)
* updated documentation GH workflows

During PR, it is now checked whether the documentation is valid by
trying to build the documentation. If this succeeds, the test
succeeds as well.

* architecture code coherency improved

Now, there is a dedicated, architecture specific `initialize()`
function. This avoids code repetition (mostly log) while allowing
for maximum flexibility.

* updated to new UEFI boot procedure handling

The UEFI system table for the runtime view is now not a global
static anymore, but returned from the `exit_uefi_boot_services`
function.

* lazily and shamelessly copied code from @phip1611

* cleaned up architectureal initialization routines

* adjusted shell log (re-added trace, removed success)

* increased source code line width to 110

* moved `qemu` module

The `qemu` module was moved from `library/helper/tests`
toml`library/helper/miscellaneous`.

* adjusted to new `exit_boot_signatures` return type

The `exit_boot_signatures` function returns two objects now,
and `efi_main` as well as `kernel_main`, and integration tests
needed to be adjusted.

* follow-up on last commit

* added `lock` module to `kernel_types`

* started working on a simple, global simple-allocator

Taken from Phillip Oppermann's BlogOS series v1, we implement
a fixed-block-size allocator and use a crate for a fallback
allocator. There is currently no `#[global_allocator]` defined,
so allocation (and even compilation) does not work.

* added doc to architecture-specific functions

All `initialization` functions not have proper documentation comments
added to the them. Moreover, the log level was adjusted for some CPU
exceptions.

* addjusted to new memory init signature

The new memory initialization routine takes the UEFI memory map. For all
parts of the kernel that rely in this feature, the invocation was
adjusted.

* simplified boot code

This commit simplifies the boot code drastically by using getting rid of
certain global (static) variables previously used for the UEFI memory
map buffer. This is now allocated by the UEFI boot services (boot time)
themselves. This allows for dropping some checks and simplifies the
code.

* started working on memory

The `kernel_heap_allocator` module now contains an (unfinished!)
implementation of a kernel heap allocator. The virtual memory module
shall now be implemted to allow for paging and proper heap allocations.

* clippy clean state

(Integration) Tests do not pass currently, on purpose. The kernel heap
allocator does not work yet, as it requires paging to be set up, which
is not yet implemented.

* make linters happy

* update EditorConfig and git attributes

* Updated community information / documentation

The in the repositry root existing files for community guidelines (and
more) (like `CONTRIBUTING.md`, `CODE_OF_CONDUCT.md`, etc.) have been
moved to the documentation and the remaining files contain the link to
the most up-to-date documentation version.

* updated documentation

The architectural overview was moved and renamed. The documentation
configuration was extended to include the new community documents.

* miscellaneous

- formatting source code
- spaces became tabs in `Justfile`
- added comunity overview content to the documentation

* updated git and formatting files

* Reintroduce `bootloader` crate (#27)

* started working on paging again

This commit does several small preparation steps:

1. Refined `kernel/Cargo.toml` (clarity) and added bootloader physical
   memory mapping
2. Moved parts of contents of the helper module for alphabetic order
3. Passing boot information now through to the memory routines

* satisfy linters

* added simple allocator and paging

This commit provides a more or less "quick and dirty" first
implementation of a page table and kernel heap allocator. This code is
not final, but works. It needs some proper cleanup, and in the future,
types that are architecture agnostic.

* added documentation checking for CI

`cargo doc --lib ...` is now executed when linting the kernel,
effectively checking whether the documentation can be properly build.

* fixed source code documentation

* improved kernel log

The kernel is now adhering to the `LOG_LEVEL` environment variable,
which overrides other, constant values set in the binary, library or
tests.

* improved timeout handler for tests

The kernel binary is now allowed to take longer than the tests, which
can now take up to 30s each. The kernel library / binary can take up to
120s.

* adjusted workspace helper (logging)

See 2a34435065e79c04eb8d775f5f2afd7114df116e

* miscellaneous small additions

* started working on memory abstractions

* make clippy happy