syncronizing onto `development` (`0.4.6`) (#33)
* update tagging convention for release tags

Instead of `vX.X.X`, we use `X.X.X`.

* updated docs and getting started section(s)

* remove unnecessary links to the Rust website

* feature: virtual memory and heap allocator (1) (#28)

* Bump github/super-linter from 4.8.7 to 4.9.0 (#29)

* feature: proper virtual memory (2/4) (#32)

* added error module and memory error types

By using a `errors` module under `kernel_types`, the kernel has gotten
easier to handle by using the `?` operator. This simplifies many tasks,
and makes the whole code easier to handle and read.

We would like to have a kernel error trait in the future that makes
using the specific error types even easier.

* added `u32` additions to addresses

* finalized heap

It turned out the "problems" of the heap were not problems at all but
due to the lazy initialization of the fixed block size allocator.

* added proper kernel `Error` trait

Implemented this trait on the error enums that are currently present.

* added functionality around page unmapping

The `deallocate_page` function itself is not yet implemented. But the
framework functions (their skeletons) that use it are there.

* submodule: libbash & doc updates & templates (#30)

* corrected documentation background color issue

* updated submodule `scripts/libbash/`

* Away from Bash, Towards Rust (#31)

* renamed `workspace_helper` => `helper`

* added kernel helper Rust code to replace Bash (1/X)

* removed `boot` workspace member

* adjusted `Justfile` and scripts

* cleaned test runner

* kernel log adjustments and misc small improvements

* added proper library to `helper`

* added binary to `helper` to get rid of Bash

* armed clippy and fired away

* small log update

* update docs and workflows

Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>